+++ model_navigator/pipelines/builders/preprocessing.py
@@ -17,11 +17,13 @@ from typing import Dict, List
 
 from model_navigator.api.config import Format
 from model_navigator.commands.base import ExecutionUnit
+from model_navigator.commands.copy.copy_model import CopyModel
 from model_navigator.commands.data_dump.samples import FetchInputModelData, FetchOutputModelData
 from model_navigator.commands.infer_metadata import InferInputMetadata, InferOutputMetadata
 from model_navigator.commands.load import LoadMetadata
 from model_navigator.configuration.common_config import CommonConfig
 from model_navigator.configuration.model.model_config import ModelConfig
+from model_navigator.frameworks import Framework
 from model_navigator.pipelines.pipeline import Pipeline
 
 
@@ -51,5 +53,10 @@ def preprocessing_builder(config: CommonConfig, models_config: Dict[Format, List
         )
     else:
         execution_units.extend([ExecutionUnit(command=LoadMetadata)])
+
+    if config.framework == Framework.ONNX:
+        execution_units.append(ExecutionUnit(command=CopyModel, model_config=models_config[Format.ONNX][0]))
+    elif config.framework == Framework.TENSORRT:
+        execution_units.append(ExecutionUnit(command=CopyModel, model_config=models_config[Format.TENSORRT][0]))
 
     return Pipeline(name="Preprocessing", execution_units=execution_units)