+++ CHANGELOG.md
@@ -22,6 +22,7 @@ limitations under the License.
 - new: Extended custom configs to pass arguments directly to export and conversion operations like `torch.onnx.export` or `polygraphy convert`
 - new: Collect GPU clock during model profiling
 - new: Add option to configure minimal trials and stabilization windows for performance verification and profiling
+- new: nav.tensorrt.optimize api added for testing and measuring peroformance of TensorRT models
 - change: Navigator package version change to 0.2.3. Custom configurations now use trt_profiles list instead single value
 
 ## 0.6.3